pyproject.toml
[project]
name = "g13"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bittty>=0.0.10",
    "hidapi>=0.15.0",
    "loguru>=0.7.3",
    "pillow>=12.0.0",
    "pynput>=1.8.1",
    "pyusb>=1.3.1",
]
